ai_service/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiolimiter==1.2.1
aiosignal==1.3.2
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
astroid==3.1.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
branca==0.8.1
build==1.4.4
cachetools==5.5.2
captcha==0.7.1
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.6
chromadb==1.5.8
click==8.3.1
cohere==5.20.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.9
django-extensions==3.2.3
djangorestframework==3.14.0
dnspython==2.7.0
docarray==0.41.0
docker==7.1.0
docutils==0.21.2
durationpy==0.10
ecdsa==0.19.1
eval_type_backport==0.3.0
executing==2.2.0
extra-streamlit-components==0.1.80
fastapi==0.121.3
fastavro==1.12.1
fastjsonschema==2.21.1
filelock==3.25.2
flatbuffers==25.12.19
folium==0.19.7
fonttools==4.58.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
googleapis-common-protos==1.72.0
greenlet==3.2.3
groq==0.4.2
grpcio==1.78.0
grpcio-health-checking==1.68.0
grpcio-reflection==1.68.0
grpcio-status==1.71.2
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface_hub==1.7.2
id==1.5.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==7.1.0
invoke==2.2.1
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jcloud==0.3
jedi==0.19.2
jina==3.34.0
jina-hubble-sdk==0.39.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keyring==25.6.0
kiwisolver==1.4.8
kubernetes==35.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.67
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.4
lazy-object-proxy==1.10.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplot==0.1.9
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistralai==1.9.11
mistune==3.1.3
mmh3==5.2.1
more-itertools==10.7.0
multidict==6.6.3
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
narwhals==1.41.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nh3==0.2.21
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.25.0
openai==1.93.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-exporter-prometheus==0.59b0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-aiohttp-client==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-grpc==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.10.18
overrides==7.7.0
packaging==26.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pillow==11.0.0
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
plotly==6.1.2
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.27.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.10.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyloco==0.0.139
pymongo==4.13.0
pyparsing==3.2.3
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2024.1
pywin32==310
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.3
pyzmq==27.0.0
readme_renderer==44.0
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==14.3.3
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==82.0.1
shellingham==1.5.4
SimpleWebSocketServer==0.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.50.0
streamlit==1.45.1
streamlit-authenticator==0.4.2
streamlit_folium==0.25.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.3
traitlets==5.14.3
twine==6.1.0
typer==0.24.1
typer-slim==0.20.0
types-python-dateutil==2.9.0.20250516
types-requests==2.32.4.20260324
types-urllib3==1.26.25.14
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
ushlex==0.99.1
uvicorn==0.23.1
voyageai==0.3.5
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.46.3
whitenoise==6.6.0
widgetsnbextension==4.0.14
wrapt==1.16.0
xyzservices==2025.4.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
